LayoutTests/sputnik/Conformance/15_Native_Objects/15.1_The_Global_Object/S15.1_A2_T1-expected.txt

S15.1_A2_T1

PASS TypeError: '[object DOMWindow]' is not a function (evaluating 'this()')

TEST COMPLETE
